+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Changelog for NeoFS Node
 - Tree service now synchronizes with container nodes in a random order (#2127)
 - Pilorama no longer tries to apply already applied operations (#2161)
 - Use `sync.Pool` in Object.PUT service (#2139)
+- Shard uses metabase for `HEAD` requests by default, not write-cache (#2167)
 
 ### Fixed
 - Open FSTree in sync mode by default (#1992)